update list table
